refactor(google): shared Drive-item tree_ops and make_rm for the docs family
gdocs/gsheets/gslides stat and unlink collapse onto
core/google/tree_ops factories (the silent readdir-populate except now
logs); their byte-identical rm wrappers bind
generic/rm_command.make_rm, which exists because the factory rm builder
calls unlink without the index these backends resolve ids through.
